move `src/stores` into `src/store`
