+++ src/app/core/catalog/cataloglayersstoresregistry.js
@@ -1,11 +1,11 @@
 /**
- * DEPRECATED: this file will be removed after v3.4 (use "stores/catalog-layers-stores.js" instead)
+ * DEPRECATED: this file will be removed after v3.4 (use "store/catalog-layers-stores.js" instead)
  */
-import CatalogLayersStoresRegistry from 'stores/catalog-layers';
+import CatalogLayersStoresRegistry from 'store/catalog-layers';
 
 /**
  * FIXME: application is broken using like the following line
  */
-// import { CatalogLayersStoresRegistry } from 'stores';
+// import { CatalogLayersStoresRegistry } from 'store';
 
 module.exports = CatalogLayersStoresRegistry;
