+++ src/app/core/editing/sessionsregistry.js
@@ -1,11 +1,11 @@
 /**
- * DEPRECATED: this file will be removed after v3.4 (use "stores/sessions.js" instead)
+ * DEPRECATED: this file will be removed after v3.4 (use "store/sessions.js" instead)
  */
-import SessionsRegistry from 'stores/sessions';
+import SessionsRegistry from 'store/sessions';
 
 /**
  * FIXME: application is broken using like the following line
  */
-// import { SessionsRegistry } from 'stores';
+// import { SessionsRegistry } from 'store';
 
 module.exports = SessionsRegistry;
