+++ src/app/core/map/maplayersstoresregistry.js
@@ -1,11 +1,11 @@
 /**
- * DEPRECATED: this file will be removed after v3.4 (use "stores/map-layers.js" instead)
+ * DEPRECATED: this file will be removed after v3.4 (use "store/map-layers.js" instead)
  */
-import MapLayersStoresRegistry from 'stores/map-layers';
+import MapLayersStoresRegistry from 'store/map-layers';
 
 /**
  * FIXME: application is broken using like the following line
  */
-// import { MapLayersStoresRegistry } from 'stores';
+// import { MapLayersStoresRegistry } from 'store';
 
 module.exports = MapLayersStoresRegistry;
