+++ src/app/gui/component/componentsregistry.js
@@ -1,12 +1,12 @@
 /**
- * DEPRECATED: this file will be removed after v3.4 (use "stores/components.js" instead)
+ * DEPRECATED: this file will be removed after v3.4 (use "store/components.js" instead)
  */
-import ComponentsRegistry from 'stores/components';
+import ComponentsRegistry from 'store/components';
 
 /**
  * FIXME: application is broken using like the following line
  */
-// import { ComponentsRegistry } from 'stores';
+// import { ComponentsRegistry } from 'store';
 
  module.exports = ComponentsRegistry;
