+++ src/app/gui/map/control/registry.js
@@ -1,12 +1,12 @@
 /**
- * DEPRECATED: this file will be removed after v3.4 (use "stores/map-controls.js" instead)
+ * DEPRECATED: this file will be removed after v3.4 (use "store/map-controls.js" instead)
  */
- import ControlsRegistry from 'stores/map-controls';
+ import ControlsRegistry from 'store/map-controls';
 
 /**
  * FIXME: application is broken using like the following line
  */
-// import { ControlsRegistry } from 'stores';
+// import { ControlsRegistry } from 'store';
 
 module.exports = ControlsRegistry;
