+++ src/store/plugins.js
